guard conditions: only try to take waitables if they are ready
This avoids doing unnecessary allocations of native ROS messages only to check
that there is no new data. This might be a good thing to do right now as with
guard_conditions there could be probably more times the waitset is ready to fire
some callbacks, so work should be minimized here especially.