extend idl generation to include interfaces for "action wrapper types"
With "action wrapper types" this means the ROS messages and services that add the `GoalId`, `TimeStamp` or other fields on top of the user defined types for the action.
The introduced methods in `IRosActionDefinition` are needed to work around the missing existential type support in C# (https://github.com/dotnet/csharplang/issues/5556).