add rosidl_core repo to ros2_dotnet_rolling.repos
In ros2/rosidl_defaults#22, the rosidl_core_generators package was split off from rosidl_defaults into a separate repo, which needed to be added.
Without this repository the dotnet code for the interfaces is not generated in the workspace.

see PR on ros2-rust: https://github.com/ros2-rust/ros2_rust/pull/251